Implementa anexos, PDF com FOP/XSL-FO, MinIO e controles de ciclo de vida

+++ backend/src/main/java/br/com/danielchipolesch/domain/services/DocumentoService.java
@@ -6,6 +6,10 @@ import br.com.danielchipolesch.application.dtos.documentoDtos.DocumentoResponseS
 import br.com.danielchipolesch.domain.builders.DocumentoBuilder;
 import br.com.danielchipolesch.domain.entities.estruturaDocumento.Documento;
 import br.com.danielchipolesch.domain.entities.estruturaDocumento.DocumentoStatusEnum;
+import br.com.danielchipolesch.domain.entities.estruturaDocumento.TipoAlteracaoEnum;
+import br.com.danielchipolesch.domain.entities.estruturaDocumento.ItemAnexoParteNormativa;
+import br.com.danielchipolesch.domain.entities.estruturaDocumento.ItemPartePreliminar;
+import br.com.danielchipolesch.domain.entities.estruturaDocumento.ItemParteFinal;
 import br.com.danielchipolesch.domain.entities.numeracaoDocumento.AssuntoBasico;
 import br.com.danielchipolesch.domain.entities.numeracaoDocumento.EspecieNormativa;
 import br.com.danielchipolesch.domain.handlers.exceptions.ResourceNotFoundException;
@@ -14,10 +18,13 @@ import br.com.danielchipolesch.domain.handlers.exceptions.enums.DocumentExceptio
 import br.com.danielchipolesch.domain.handlers.exceptions.StatusCannotBeUpdatedException;
 import br.com.danielchipolesch.domain.handlers.exceptions.enums.DocumentationTypeException;
 import br.com.danielchipolesch.domain.mappers.DocumentoMapper;
+import br.com.danielchipolesch.domain.entities.estruturaDocumento.Anexo;
+import br.com.danielchipolesch.infrastructure.repositories.AnexoRepository;
 import br.com.danielchipolesch.infrastructure.repositories.AssuntoBasicoRepository;
 import br.com.danielchipolesch.infrastructure.repositories.DocumentoRepository;
 import br.com.danielchipolesch.infrastructure.repositories.EspecieNormativaRepository;
 import br.com.danielchipolesch.infrastructure.repositories.ItemAnexoParteNormativaRepository;
+import br.com.danielchipolesch.infrastructure.repositories.DocumentoHistoricoRepository;
 import br.com.danielchipolesch.infrastructure.repositories.ItemParteFinalRepository;
 import br.com.danielchipolesch.infrastructure.repositories.ItemPartePreliminarRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -50,6 +57,15 @@ public class DocumentoService {
     @Autowired
     ItemParteFinalRepository itemParteFinalRepository;
 
+    @Autowired
+    DocumentoHistoricoRepository documentoHistoricoRepository;
+
+    @Autowired
+    DocumentoHistoricoService documentoHistoricoService;
+
+    @Autowired
+    AnexoRepository anexoRepository;
+
 
     @Transactional
     public DocumentoResponseSemAnexoTextualDto create(DocumentoRequestCreateDto request) throws RuntimeException {
@@ -67,7 +83,10 @@ public class DocumentoService {
                 .documentoStatus(DocumentoStatusEnum.RASCUNHO)
                 .build();
 
-        return DocumentoMapper.documentoToDocumentoSemAnexoTextualResponseDto(documentoRepository.save(documento));
+        Documento salvo = documentoRepository.save(documento);
+        documentoHistoricoService.registrar(salvo, TipoAlteracaoEnum.CRIACAO,
+                "Documento criado", null, DocumentoStatusEnum.RASCUNHO);
+        return DocumentoMapper.documentoToDocumentoSemAnexoTextualResponseDto(salvo);
     }
 
     public Documento getById(Long id) throws RuntimeException{
@@ -105,25 +124,44 @@ public class DocumentoService {
             throw new StatusCannotBeUpdatedException(DocumentException.CANNOT_BE_UPDATED.getMessage());
         }
 
+        boolean tituloAlterado = !document.getTituloDocumento().equals(request.getTituloDocumento());
         document.setTituloDocumento(request.getTituloDocumento());
-        return DocumentoMapper.documentoToDocumentoSemAnexoTextualResponseDto(documentoRepository.save(document));
+        if (request.getNumeroSecundario() != null) {
+            document.setNumeroSecundario(request.getNumeroSecundario());
+        }
+        Documento atualizado = documentoRepository.save(document);
+        if (tituloAlterado) {
+            documentoHistoricoService.registrar(atualizado, TipoAlteracaoEnum.ALTERACAO_METADADOS,
+                    "Título atualizado", null, null);
+        }
+        return DocumentoMapper.documentoToDocumentoSemAnexoTextualResponseDto(atualizado);
     }
 
     @Transactional
     public DocumentoResponseSemAnexoTextualDto delete(Long id) throws RuntimeException {
         Documento document = documentoRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException(DocumentException.NOT_FOUND.getMessage()));
+
+        DocumentoStatusEnum status = document.getDocumentoStatus();
+        if (status != DocumentoStatusEnum.RASCUNHO && status != DocumentoStatusEnum.MINUTA) {
+            throw new StatusCannotBeUpdatedException(DocumentException.CANNOT_BE_DELETED.getMessage());
+        }
+
+        documentoHistoricoRepository.deleteAllByDocumentoId(id);
         itemPartePreliminarRepository.deleteAllByDocumentoId(id);
         itemAnexoParteNormativaRepository.nullifyParentsForDocument(id);
         itemAnexoParteNormativaRepository.deleteAllByDocumentoId(id);
         itemParteFinalRepository.deleteAllByDocumentoId(id);
+        anexoRepository.deleteAllByDocumentoId(id);
         documentoRepository.delete(document);
         return DocumentoMapper.documentoToDocumentoSemAnexoTextualResponseDto(document);
     }
 
+    @Transactional
     public DocumentoResponseSemAnexoTextualDto clone(Long id) throws RuntimeException {
 
-        Documento documentOld = documentoRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(DocumentException.NOT_FOUND.getMessage()));
+        Documento documentOld = documentoRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(DocumentException.NOT_FOUND.getMessage()));
 
         var secondaryNumber = this.calculateSecondaryNumber(documentOld.getEspecieNormativa(), documentOld.getAssuntoBasico());
 
@@ -135,7 +173,66 @@ public class DocumentoService {
                 .documentoStatus(DocumentoStatusEnum.RASCUNHO)
                 .build();
 
-        return DocumentoMapper.documentoToDocumentoSemAnexoTextualResponseDto(documentoRepository.save(documentNew));
+        documentOld.setQtdReplicas(documentOld.getQtdReplicas() + 1);
+        documentoRepository.save(documentOld);
+
+        Documento clonado = documentoRepository.save(documentNew);
+
+        for (ItemPartePreliminar orig : itemPartePreliminarRepository.findByDocumentoIdOrderByElementOrderAsc(id)) {
+            ItemPartePreliminar copia = new ItemPartePreliminar();
+            copia.setDocumento(clonado);
+            copia.setTipo(orig.getTipo());
+            copia.setElementOrder(orig.getElementOrder());
+            copia.setTitulo(orig.getTitulo());
+            copia.setConteudo(orig.getConteudo());
+            copia.setFullTextContent(orig.getFullTextContent());
+            itemPartePreliminarRepository.save(copia);
+        }
+
+        for (ItemAnexoParteNormativa root : itemAnexoParteNormativaRepository.findRootItemsByDocumentoId(id)) {
+            clonarNormItem(root, clonado, null);
+        }
+
+        for (ItemParteFinal orig : itemParteFinalRepository.findByDocumentoIdOrderByElementOrderAsc(id)) {
+            ItemParteFinal copia = new ItemParteFinal();
+            copia.setDocumento(clonado);
+            copia.setTipo(orig.getTipo());
+            copia.setElementOrder(orig.getElementOrder());
+            copia.setTitulo(orig.getTitulo());
+            copia.setConteudo(orig.getConteudo());
+            copia.setFullTextContent(orig.getFullTextContent());
+            itemParteFinalRepository.save(copia);
+        }
+
+        for (Anexo orig : anexoRepository.findByDocumentoIdOrderByOrdemAsc(id)) {
+            Anexo copia = new Anexo();
+            copia.setDocumento(clonado);
+            copia.setTitulo(orig.getTitulo());
+            copia.setUrlImagem(orig.getUrlImagem());
+            copia.setOrdem(orig.getOrdem());
+            anexoRepository.save(copia);
+        }
+
+        documentoHistoricoService.registrar(clonado, TipoAlteracaoEnum.CLONAGEM,
+                "Clonado do documento #" + id, null, DocumentoStatusEnum.RASCUNHO);
+        return DocumentoMapper.documentoToDocumentoSemAnexoTextualResponseDto(clonado);
+    }
+
+    private void clonarNormItem(ItemAnexoParteNormativa original, Documento novoDoc, ItemAnexoParteNormativa novoParent) {
+        ItemAnexoParteNormativa copia = new ItemAnexoParteNormativa();
+        copia.setDocumento(novoDoc);
+        copia.setParent(novoParent);
+        copia.setTipo(original.getTipo());
+        copia.setElementOrder(original.getElementOrder());
+        copia.setTitulo(original.getTitulo());
+        copia.setConteudo(original.getConteudo());
+        copia.setFullTextContent(original.getFullTextContent());
+        ItemAnexoParteNormativa salva = itemAnexoParteNormativaRepository.save(copia);
+        if (original.getChildren() != null) {
+            for (ItemAnexoParteNormativa filho : original.getChildren()) {
+                clonarNormItem(filho, novoDoc, salva);
+            }
+        }
     }
 
     private Integer calculateSecondaryNumber(EspecieNormativa especieNormativa, AssuntoBasico assuntoBasico){
